Move `Transaction` to its own module
