+++ crates/core/src/types/mod.rs
@@ -1,7 +1,9 @@
 mod account;
 mod block;
 mod genesis;
+mod transaction;
 
 pub use account::*;
 pub use block::*;
 pub use genesis::*;
+pub use transaction::*;
